fixed config issues

+++ knowledge_distillation/src/main/class_transformers_distillation_main.py
@@ -506,7 +506,7 @@ if __name__ == '__main__':
         # student_model_name="meta-llama/Llama-3.2-1B-Instruct",  # Changed model
         student_model_name="aspenita/llama-3-sqlcoder-8b-AWQ",  # Changed model
         student_model_torch_dtype="float16",
-        teacher_model_torch_dtype="float16",
+        teacher_model_torch_dtype="float32",
         distillation_type="black_box",  # Using combined distillation
         temperature=2.0,
         alpha=0.5,
